requirements.txt
cryptography==45.0.6
python-dotenv==1.1.1
Requests==2.32.4
selenium==4.34.2
google-genai